Move play link or playlist to new play youtube location
