+++ src/music/index.ts
@@ -186,170 +186,6 @@ export const play = (message: Message, song: SongShape) => {
     });
 };
 
-const addTrackToPlaylist = async (
-  message: Message,
-  title: string,
-  url: string,
-  trackVolume: number,
-) => {
-  const voiceChannel = message.member?.voice.channel;
-  if (!voiceChannel) {
-    reactWithEmoji.failed(message);
-    logger.log({
-      level: 'error',
-      message: `No voice channel while adding track to playlist.- ${title}`,
-    });
-    return;
-  }
-  const song = {
-    id: uuidv4(),
-    title,
-    url,
-    volume: trackVolume,
-  };
-  const playlist = getPlaylist(message, defaultPlaylistName);
-  if (!playlist) {
-    const newPlaylist = await createPlaylist(
-      message,
-      defaultPlaylistName,
-      'adding a track to the playlist',
-    );
-    try {
-      if (isNull(newPlaylist)) {
-        throw new Error('No playlist found.');
-      }
-      const connection = await voiceChannel.join();
-      setPlaylist(message, defaultPlaylistName, {
-        ...newPlaylist,
-        connection,
-        currentSong: song,
-        songs: [song],
-      });
-      play(message, song);
-    } catch (error) {
-      reactWithEmoji.failed(message);
-      logger.log({
-        level: 'error',
-        message: `Error occurred while creating a new playlist: ${error.message}`,
-      });
-    }
-    return;
-  }
-  playlist.isWriteLocked = true;
-  playlist.songs.push(song);
-  setPlaylist(message, defaultPlaylistName, playlist);
-  if (
-    !playlist?.currentSong?.id ||
-    playlist?.currentSong?.id === songScaffold.id
-  ) {
-    // Need to check if playlist.connection = null
-    if (isNil(playlist.connection)) {
-      const connection = await voiceChannel.join();
-      playlist.connection = connection;
-      setPlaylist(message, defaultPlaylistName, playlist);
-    }
-    play(message, song);
-  }
-  const [previousSong, currentSong, nextSong] = dryRunTraversePlaylistByStep(
-    playlist,
-    1,
-  );
-  playlist.previousSong = previousSong;
-  playlist.currentSong = currentSong;
-  playlist.nextSong = nextSong;
-  playlist.isWriteLocked = false;
-  setPlaylist(message, defaultPlaylistName, playlist);
-};
-
-export const playAndOrAddYoutubeToPlaylist = async (message: Message) => {
-  reactWithEmoji.received(message);
-  if (!message.guild?.id) {
-    reactWithEmoji.failed(message);
-    logger.log({
-      level: 'error',
-      message: `No guild ID found while attempting to play and or add an existing YouTube link.`,
-    });
-    return;
-  }
-
-  const voiceChannel = message.member?.voice.channel;
-  if (!voiceChannel) {
-    reactWithEmoji.failed(message);
-    return message.channel.send(
-      "I'm not gonna play for no one. Someone get into a voice channel first.",
-    );
-  }
-  if (!message?.client?.user) {
-    reactWithEmoji.failed(message);
-    logger.log({
-      level: 'error',
-      message: `No user found while attempting to play and or add an existing YouTube link.`,
-    });
-    return;
-  }
-  const permissions = voiceChannel.permissionsFor(message.client.user);
-  if (!permissions?.has('CONNECT') || !permissions.has('SPEAK')) {
-    reactWithEmoji.failed(message);
-    return message.channel.send(
-      'Give me permissions for connecting and speaking in the voice channel, then we can party.',
-    );
-  }
-
-  const {
-    volume,
-    maxAllowableReached,
-    link,
-    playlistId,
-  } = getYoutubeLinkAndVolFromRequest(message.content, maxAllowableVolume);
-
-  if (link === '#') {
-    reactWithEmoji.failed(message);
-    return message.channel.send("...I can't play _that_.");
-  }
-
-  if (maxAllowableReached) {
-    message.channel.send(
-      `._shakes his head_ I won't play songs louder than a level of **${maxAllowableReached}**.`,
-    );
-  }
-
-  if (playlistId !== '-') {
-    // YouTube playlist
-    try {
-      const playlistIdIsValid = ytpl.validateID(playlistId);
-      if (playlistIdIsValid) {
-        const youtubePlaylist = await ytpl(playlistId);
-        const numberOfTracks = youtubePlaylist.items.length;
-        for (const track of youtubePlaylist.items) {
-          await addTrackToPlaylist(
-            message,
-            track.title,
-            track.shortUrl,
-            volume,
-          );
-        }
-        reactWithEmoji.succeeded(message);
-        return message.channel.send(
-          `_nods and adds_ **${numberOfTracks}** _tracks with volume at_ **${volume}** _to the list._`,
-        );
-      }
-    } catch (error) {}
-  }
-
-  // Single track
-  const songInfo = await ytdl.getInfo(link);
-  await addTrackToPlaylist(
-    message,
-    songInfo.videoDetails.title,
-    songInfo.videoDetails.video_url,
-    volume,
-  );
-  reactWithEmoji.succeeded(message);
-  return message.channel.send(
-    `_nods and adds_ **${songInfo.videoDetails.title}** with volume at **${volume}** _to the list._`,
-  );
-};
-
 export const joinVoiceChannel = async (message: Message) => {
   reactWithEmoji.received(message);
   let playlist = getPlaylist(message, defaultPlaylistName);
