+++ src/social/index.ts
@@ -19,7 +19,6 @@ import {
   loopOffPrefixCommands,
   loopPlaylistPrefixCommands,
   loopTrackPrefixCommands,
-  playYoutubeLinkPrefixCommands,
 } from '../music/constants';
 import {
   playExistingTrackMandTrackPrefixCommands,
@@ -28,6 +27,7 @@ import {
 import { setSongVolPrefixCommands } from '../music/volume/constants';
 import { showPlaylistPrefixCommands } from '../music/list/constants';
 import { removeTrackPrefixCommands } from '../music/rm/constants';
+import { playYoutubeLinkPrefixCommands } from '../music/play/youtube/constants';
 
 const botPrefix = BOT_PREFIX;
